requirements.txt
Flask==3.0.3
openai>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.1
